requirements.txt
torch>=2.1.2
torchvision
tqdm
datasets
scikit-learn==1.5.2
transformers
timm
accelerate
sentencepiece
protobuf
pymp-pypi
matplotlib
seaborn
jupyter
loguru
torchinfo
torchmetrics
wandb
deepspeed
av
cca-zoo==2.5.0
faiss-cpu
